+++ carpetmodSrc/carpet/utils/portalcalculator/PortalSilentSearcher.java
@@ -1,32 +1,26 @@
 package carpet.utils.portalcalculator;
 
 import carpet.utils.Messenger;
-import it.unimi.dsi.fastutil.longs.Long2ObjectMap;
-import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
+import carpet.utils.SilentChunkReader;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.ChunkPos;
 import net.minecraft.util.math.MathHelper;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.DimensionType;
 import net.minecraft.world.WorldServer;
 import net.minecraft.world.border.WorldBorder;
-import net.minecraft.world.chunk.Chunk;
-import net.minecraft.world.gen.ChunkProviderServer;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.io.IOException;
 import java.util.Objects;
 
 /**
  * Runs portal searching and matching without making chunks loaded.
  */
-public class PortalSilentSearcher implements Runnable
-{
+public class PortalSilentSearcher implements Runnable {
     private static final double INTER_DIM_RATE = 8.0;
     private static final double BORDER_WIDTH = 16.0;
     private static final int BORDER_POS = 29999872;
@@ -44,13 +38,13 @@ public class PortalSilentSearcher implements Runnable
      * Stores only the "valid" "lowest" portal block in portal patterns
      */
     private final PortalMegaCache portalBottomCache = new PortalMegaCache();
-    private final Long2ObjectMap<Chunk> chunkCache = new Long2ObjectOpenHashMap<>(289);
     private final MinecraftServer server;
     private final Vec3d posTarget;
     private final EnumTargetDirection direction;
     private final EnumTargetArea area;
     private final DimensionType dimension;
     private WorldServer world = null;
+    private SilentChunkReader reader = null;
     private BlockPos posCenter = null;
     private PortalPattern patternCenter = null;
     private PortalPattern patternResult = null;
@@ -58,8 +52,7 @@ public class PortalSilentSearcher implements Runnable
     private boolean successful = false;
     private double distSqCache = 0.0;
 
-    public PortalSilentSearcher(MinecraftServer server, Vec3d posTarget, DimensionType dimension, EnumTargetDirection direction, EnumTargetArea area)
-    {
+    public PortalSilentSearcher(MinecraftServer server, Vec3d posTarget, DimensionType dimension, EnumTargetDirection direction, EnumTargetArea area) {
         this.server = server;
         this.posTarget = posTarget;
         this.direction = direction;
@@ -67,8 +60,7 @@ public class PortalSilentSearcher implements Runnable
         this.dimension = dimension;
     }
 
-    public boolean isSuccessful()
-    {
+    public boolean isSuccessful() {
         return successful;
     }
 
@@ -76,88 +68,25 @@ public class PortalSilentSearcher implements Runnable
      * {@link net.minecraft.block.BlockPortal#createPatternHelper}
      */
     @Nullable
-    private PortalPattern getParentPattern(BlockPos blockPos)
-    {
-        IBlockState blockState = getBlockStateSilent(blockPos);
-        if (blockState == null)
-        {
+    private PortalPattern getParentPattern(BlockPos blockPos) {
+        IBlockState blockState = reader.getBlockState(blockPos);
+        if (blockState == null) {
             return null;
         }
-        MutablePortalPattern pattern = new MutablePortalPattern(this, blockPos, EnumFacing.Axis.X);
-        if (!pattern.isValid())
-        {
-            pattern = new MutablePortalPattern(this, blockPos, EnumFacing.Axis.Z);
+        MutablePortalPattern pattern = new MutablePortalPattern(reader, blockPos, EnumFacing.Axis.X);
+        if (!pattern.isValid()) {
+            pattern = new MutablePortalPattern(reader, blockPos, EnumFacing.Axis.Z);
         }
         return pattern.isValid() ? pattern.toImmutable() : new PortalPattern(blockPos, blockPos);
     }
 
-    @Nullable
-    public IBlockState getBlockStateSilent(BlockPos pos)
-    {
-        try
-        {
-            return getBlockStateSilent(pos.getX(), pos.getY(), pos.getZ());
-        }
-        catch (NullPointerException e)
-        {
-            return null;
-        }
-    }
-
     @Nonnull
-    public IBlockState getBlockStateSilent(int x, int y, int z) throws NullPointerException
-    {
-        if (y < 0 || y >= 256)
-        {
-            return Objects.requireNonNull(Blocks.AIR).getDefaultState();
-        }
-        else
-        {
-            Chunk chunk = getChunkSilent(x >> 4, z >> 4);
-            return Objects.requireNonNull(chunk).getBlockState(x, y, z);
-        }
-    }
-
-    @Nullable
-    private Chunk getChunkSilent(int x, int z)
-    {
-        long index = ChunkPos.asLong(x, z);
-        if (chunkCache.containsKey(index))
-        {
-            return chunkCache.get(index);
-        }
-        Chunk chunk = null;
-        ChunkProviderServer provider = world.getChunkProvider();
-        if (provider.chunkExists(x, z))
-        {
-            chunk = provider.loadedChunks.get(ChunkPos.asLong(x, z));
-        }
-        else if (provider.chunkLoader.isChunkGeneratedAt(x, z))
-        {
-            try
-            {
-                chunk = provider.chunkLoader.loadChunk_silent(world, x, z);
-            }
-            catch (IOException ignored)
-            {
-            }
-        }
-        if (chunk != null)
-        {
-            chunkCache.put(index, chunk);
-        }
-        return chunk;
-    }
-
-    @Nonnull
-    private BlockPos clampTeleportDestination(@Nonnull Vec3d posTeleported) throws NullPointerException
-    {
+    private BlockPos clampTeleportDestination(@Nonnull Vec3d posTeleported) throws NullPointerException {
         return clampTeleportDestination(posTeleported.x, posTeleported.y, posTeleported.z);
     }
 
     @Nonnull
-    private BlockPos clampTeleportDestination(double x, double y, double z) throws NullPointerException
-    {
+    private BlockPos clampTeleportDestination(double x, double y, double z) throws NullPointerException {
         // copied from vanilla codes
         WorldBorder borderObj = Objects.requireNonNull(world).getWorldBorder();
         x = MathHelper.clamp(x, borderObj.minX() + BORDER_WIDTH, borderObj.maxX() - BORDER_WIDTH);
@@ -167,16 +96,13 @@ public class PortalSilentSearcher implements Runnable
         return new BlockPos(x, y, z);
     }
 
-    private void initFields() throws NullPointerException
-    {
+    private void initFields() throws NullPointerException {
         double x = posTarget.x;
         double y = posTarget.y;
         double z = posTarget.z;
-        switch (direction)
-        {
+        switch (direction) {
             case FROM:
-                switch (dimension)
-                {
+                switch (dimension) {
                     case NETHER:
                         x *= INTER_DIM_RATE;
                         z *= INTER_DIM_RATE;
@@ -193,6 +119,7 @@ public class PortalSilentSearcher implements Runnable
                 world = server.getWorld(dimension.getId());
                 break;
         }
+        reader = world.silentChunkReader;
         posCenter = clampTeleportDestination(x, y, z);
         initialized = true;
     }
@@ -201,29 +128,23 @@ public class PortalSilentSearcher implements Runnable
      * Simulates vanilla searching
      */
     @Nonnull
-    private BlockPos findPointDestination(BlockPos posOrigin)
-    {
+    private BlockPos findPointDestination(BlockPos posOrigin) {
         final int actualLimit = world.getActualHeight() - 1;
         BlockPos.PooledMutableBlockPos posResult = BlockPos.PooledMutableBlockPos.retain();
         BlockPos.PooledMutableBlockPos posPortal = BlockPos.PooledMutableBlockPos.retain();
         double distSqMin = -1.0;
-        for (int bx = -128; bx <= 128; ++bx)
-        {
+        for (int bx = -128; bx <= 128; ++bx) {
             int xDetect = posOrigin.getX() + bx;
-            for (int bz = -128; bz <= 128; ++bz)
-            {
+            for (int bz = -128; bz <= 128; ++bz) {
                 int zDetect = posOrigin.getZ() + bz;
                 posPortal.setPos(xDetect, 0, zDetect);
-                for (int yDetect = actualLimit; yDetect >= 0; --yDetect)
-                {
-                    IBlockState stateToDetect = getBlockStateSilent(xDetect, yDetect, zDetect);
-                    if (stateToDetect.getBlock() == Blocks.PORTAL)
-                    {
+                for (int yDetect = actualLimit; yDetect >= 0; --yDetect) {
+                    IBlockState stateToDetect = reader.getBlockState(xDetect, yDetect, zDetect);
+                    if (stateToDetect.getBlock() == Blocks.PORTAL) {
                         Messenger.print_server_message(server, String.format("Detected PORTAL block at %s", new BlockPos(xDetect, yDetect, zDetect)));
                         // find the lowest portal block in current portal pattern to detect
                         int yBottom = yDetect - 1;
-                        while (getBlockStateSilent(xDetect, yBottom, zDetect).getBlock() == Blocks.PORTAL)
-                        {
+                        while (reader.getBlockState(xDetect, yBottom, zDetect).getBlock() == Blocks.PORTAL) {
                             Messenger.print_server_message(server, String.format("Detected PORTAL block at %s", new BlockPos(xDetect, yBottom, zDetect)));
                             --yBottom;
                         }
@@ -231,8 +152,7 @@ public class PortalSilentSearcher implements Runnable
                         posPortal.setY(yBottom + 1);
                         double distSqTemp = posPortal.distanceSq(posOrigin);
                         Messenger.print_server_message(server, String.format("portal at %s, dist = %.1f", posPortal, Math.sqrt(distSqTemp)));
-                        if (distSqMin < 0.0 || distSqTemp < distSqMin)
-                        {
+                        if (distSqMin < 0.0 || distSqTemp < distSqMin) {
                             Messenger.print_server_message(server, String.format("closer portal! %.1f -> %.1f", Math.sqrt(distSqMin), Math.sqrt(distSqTemp)));
                             distSqMin = distSqTemp;
                             posResult.setPos(posPortal);
@@ -246,24 +166,17 @@ public class PortalSilentSearcher implements Runnable
     }
 
     @Override
-    public void run()
-    {
-        try
-        {
+    public void run() {
+        try {
             successful = false;
             initFields();
-            if (area == EnumTargetArea.RANGE)
-            {
+            if (area == EnumTargetArea.RANGE) {
                 patternCenter = getParentPattern(posCenter);
-            }
-            else
-            {
+            } else {
                 patternCenter = new PortalPattern(posCenter, posCenter);
             }
-            if (direction == EnumTargetDirection.FROM)
-            {
-                if (area == EnumTargetArea.POINT)
-                {
+            if (direction == EnumTargetDirection.FROM) {
+                if (area == EnumTargetArea.POINT) {
                     BlockPos posDest = findPointDestination(posCenter);
                     PortalPattern patternDest = getParentPattern(posDest);
                     Messenger.print_server_message(server, String.format("Destination Block: %s, dist: %.1f", posDest, Math.sqrt(distSqCache)));
@@ -271,9 +184,7 @@ public class PortalSilentSearcher implements Runnable
                 }
             }
             successful = true;
-        }
-        catch (Exception e)
-        {
+        } catch (Exception e) {
             Messenger.print_server_message(server, "Exceptions in \"/portal\"");
             e.printStackTrace();
         }